backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
pydantic-settings>=2.0.0

# ML & Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
git+https://github.com/openai/CLIP.git@main

# Vector Database & Similarity Search
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
imagehash>=4.3.1

# Async & Caching
aiofiles>=23.0.0
aiohttp>=3.9.0
httpx>=0.25.0
diskcache>=5.6.3

# Web Scraping & Data
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
google-search-results>=2.4.2   # SerpAPI Python client
ddgs>=9.0.0                   # DuckDuckGo search (no API key needed, replaces duckduckgo-search)


# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL

# Caching & Session Management
redis>=5.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0

# Logging & Monitoring
sentry-sdk>=1.38.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
python-dotenv>=1.0.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0